pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cloudflare-ddns-updater"
version = "0.1.0"
description = "Simple Cloudflare DDNS updater"
authors = [{name = "Your Name"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "requests>=2.31,<3",
  "python-dotenv>=1,<2",
]

[project.scripts]
cloudflare-ddns = "ddns.__main__:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["ddns"]
